+++ server/bedfellow-api/src/controllers/mod.rs
@@ -1,0 +1,2 @@
+pub mod spotify_controller;
+pub mod whosampled_controller;
