+++ server/bedfellow-api/src/run_python_scraper.sh
